pyproject.toml
[project]
name = "flight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-adk>=1.12.0",
    "google-genai>=1.31.0",
    "uvicorn>=0.35.0",
]
